Deployment ready 2021-2022 updates
