--- src/pages/OLD members/assets/data.json
@@ -1,110 +1,0 @@
-{
-	"main": {
-		"heading": "2019-2020 UAV Team",
-		"subheading": "Interested. Active. Engaged. All of these describe the members of our team, who love working together and learning new things.",
-		"caption": [
-			"Pictured left to right, top row: Niko Economos, Daniel Stefanescu, Ganesh Nanduru, Liam West, Ronith Reddy, Ron Nachum, Aaryan Dave.",
-			"Second row: James Huang, Caleb Arulandu, Kevin Gu, Patrick Zhang, Jason Chen, Jack Blair, Cyril Sharma, Alexander Black.",
-			"Third row: Jacob Consalvi, Aryan Chinnala, Lohit Balakumar, Srikar Gouru, Ari Bobesh, Audhav, Durai, Jason Yi, Ankit Khandelwal.",
-			"Bottom row: Jon Paolo Recta, Max Zhang, Jack Williamson, Suruchi Sen, Hari Shanmugaraja, Marvin Fu."
-		]
-	},
-	"officers": {
-		"heading": "Officers",
-		"entries": [
-			{
-				"name": "Srikar Gouru",
-				"position": "President",
-				"image": {
-					"src": "srikar.jpg",
-					"alt": "Srikar Gouru; President"
-				}
-			},
-			{
-				"name": "Ganesh Nanduru",
-				"position": "Vice President",
-				"image": {
-					"src": "ganesh.jpg",
-					"alt": "Ganesh Nanduru; Vice President"
-				}
-			},
-			{
-				"name": "Niko Economos",
-				"position": "Secretary",
-				"image": {
-					"src": "niko.jpg",
-					"alt": "Niko Economos; Secretary"
-				}
-			},
-			{
-				"name": "Liam West",
-				"position": "Treasurer",
-				"image": {
-					"src": "liam.jpg",
-					"alt": "Liam West; Treasurer"
-				}
-			},
-			{
-				"name": "Jason Klein",
-				"position": "Webmaster",
-				"image": {
-					"src": "jasonk.jpg",
-					"alt": "Jason Klein; Webmaster"
-				}
-			},
-			{
-				"name": "Daniel Stefanescu",
-				"position": "Publicist",
-				"image": {
-					"src": "daniel.jpg",
-					"alt": "Daniel Stefanescu; Publicist"
-				}
-			}
-		]
-	},
-	"leads": {
-		"heading": "Team Leads",
-		"entries": [
-			{
-				"name": "Alex Black",
-				"position": "CV Team",
-				"image": {
-					"src": "alex.jpg",
-					"alt": "Alex Black; CV Team lead"
-				}
-			},
-			{
-				"name": "Ari Bobesh",
-				"position": "UGV Team",
-				"image": {
-					"src": "ari.jpg",
-					"alt": "Ari Bobesh; UGV Team lead"
-				}
-			},
-			{
-				"name": "Aryan Chinnala",
-				"position": "GCS Team",
-				"image": {
-					"src": "aryan.jpg",
-					"alt": "Aryan Chinnala; GCS Team lead"
-				}
-			},
-			{
-				"name": "Ganesh Nanduru",
-				"position": "Programming",
-				"image": {
-					"src": "ganesh.jpg",
-					"alt": "Ganesh Nanduru; Programming lead"
-				}
-			},
-			{
-				"name": "Kevin Gu",
-				"position": "CAD Team",
-				"image": {
-					"src": "kevin.jpg",
-					"alt": "Kevin Gu; CAD Team lead"
-				}
-			}
-		]
-	}
-}
