--- src/pages/OLD members/components/Big Cards/Content.js
@@ -1,23 +1,0 @@
-import React from "react"
-import styled from "styled-components"
-
-import Card from "./BigCard"
-
-const Content = ({ entries, bg_color }) => (
-	<Container>
-		{entries.map((lead, i) => (
-			<Card key={i} data={lead} bg_color={bg_color} />
-		))}
-	</Container>
-)
-
-export default Content
-
-const Container = styled.div`
-	--rows: unset;
-	display: grid;
-	grid-template-columns: repeat(auto-fill, minmax(14rem, 1fr));
-	grid-auto-rows: min-content;
-	gap: 1rem;
-	grid-area: content;
-`
