+++ src/pages/members/components/Big Cards/Officers.js
@@ -4,13 +4,14 @@ import styled from "styled-components"
 import { StyledHeading } from "theme/Styles"
 import { officers } from "pages/members/assets/data.json"
 
+import Container from "./Container"
 import Content from "./Content"
 
 const Officers = () => (
-	<>
+	<Container as="section" area="officers">
 		<Heading>{officers.heading}</Heading>
 		<Content entries={officers.entries} />
-	</>
+	</Container>
 )
 
 const Heading = styled(StyledHeading)`
